Add inplace kwarg to ionize & recombine in Particle
Prior to this commit, I had Particle.ionize and Particle.recombine
both change the instance of Particle that these methods were called
from to the ionized/recombined Particle, and return the new
ionized/recombined Particle as well.  This commit adds the `inplace`
keyword argument that allows users to specify that the Particle should
or should not be changed to the new Particle or returned.  The name
`inplace` is the convention used in Pandas.

Co-authored by: Dominik Stańczak <stanczakdominik@gmail.com>