Move information in INSTALL.md to docs/install.rst
This way we will have our installation instructions showing up in our
documentation guide (including in searches) rather than linking back
to our GitHub repository.

In INSTALL.md, the link I put to the installation instructions are to
the `latest` rather than `stable` version of our docs.  At the moment
the stable version of our docs does not contain that page.

We will need to update links on our webpage and elsewhere in the docs
and repository from INSTALL.md to the installation instructions in our
docs.