Reorganize docs on installing PlasmaPy for development
The directions on how to install a PlasmaPy for development have been
in code_guide.rst, which also contained code style guidelines.  This
commit puts the development installation instructions in a separate
file.

For the time being, I put the Cython instructions together with the
installation instructions.  I am wondering if we should create a new
file for the Cython instructions too.