Update user installation doc page
I have been wavering between whether or not the user installation
instructions should be on the same page as the contributor
installation instructions.  Including all of the installation
instructions in one place may make it seem overwhelming to new users,
so right now I am thinking that it would make the most sense to direct
docs/install.rst to new users and docs/development/install_dev.rst to
new contributors.

The things that still need to be added to this page are instructions
on creating a conda environment to install PlasmaPy in, and
instructions on how to build PlasmaPy.

I don't think I have yet included all of the prerequisites for
building PlasmaPy.

Much of the reorganization and layout is inspired by Astropy's
installation instructions.