Change "# coveralls: ignore" to "# coverage: ignore"
This renames the comment that we use to specify that a line of code
should not be counted in coverage reports (e.g., these lines will not
be marked as tested or untested).  We now use Codecov instead of
Coveralls, and this commit makes it more general that we ignore these
lines for coverage tests rather than one of the above specific
platforms.