Update test documentation (#542)
This pull request updates the testing guide in our development
guide since it needed to be updated to reflect that we are using
Codecov now intead of Coveralls, as well as a bunch of other
things.

I prepended a motivation section since the plasma physics
community doesn't always recognize the importance of testing.
I split the other sections up into how to run tests, write tests,
and use code coverage reports.

I added a discussion about using `python setup.py test`.  I'm
wondering if it would be helpful to add some more information
later on about the integration with setuptools, though maybe that
would be best described in the code itself.  

I expanded the discussion about testing warnings and exceptions,
and tried to change function names in examples to say more
explicitly what the function does (even though the function names
are now kind of boring).

I provided a more thorough discussion of the pytest helpers that
I contributed a while back.  I still think of this feature as
experimental, though I personally really like it since it makes
it much easier to write tests that compare Quantity objects to
within a tolerance and to check warnings and exceptions.

I kept the discussion on fixtures brief as suggested, though we
might want to put an example of the usage we might expect for
plasma physics in the documentation later on.

I updated the discussion on using coverage reports with Codecov,
and described how to generate these reports locally.

I added mypy cache directories to .gitignore in case we ever end
up using mypy, and updated our contributing guide.