Have mass_energy use mass of nucleon instead
The mass_energy function and Particle.mass_energy had just been
calculating the mass energy for isotopes based on the total mass, with
nuclear_mass_energy and Particle.nuclear_mass_energy being what
calculated the mass energy of the nucleus of an isotope.

The mass_energy function will generally be used for nuclear reactions,
where the nuclear mass energy is the only important quantity.  This
commit removes nuclear_mass_energy and Particle.nuclear_mass_energy
and makes mass_energy and Particle.mass_energy calculate the mass
energy based off of the mass of the nucleus of an isotope or nucleon.