Add Particle.__hash__
I tried to make a dict with Particle instances as keys, but was not
able to because __hash__ was not defined.  I took the hash of
Particle(...).__repr__(), at least temporarily.  I did not add a
docstring to __hash__ because (1) the implementation is pretty
straightforward, and (2) I don't understand what is going on well
enough to write one.  Insert appropriate cat emoji here.