Stop ignoring __repr__ in coverage tests
In general, it is more helpful to have untested lines be marked as
untested rather than ignored.  It is often unimportant to test __str__
and __repr__, but there are occasions where tests of them may be
worthwhile.  For example, I am using __repr__ in IonizationStates to
print diagnostic information, and it is useful to test that it does
not raise any exceptions at the very least.