Improve IonizationStates.__setitem__ and tests
I reworked __setitem__ because it had a whole bunch of nested
try/except and if statements that made the flow hard to parse.  I also
added a few tests and fixed a few bugs.

I also started using pytest.fail.Exception as the exception to be
raised in test failures since that seems more appropriate than
something like AtomicError.  We still use AssertionError a lot too,
which is still appropriate.