Update inertial_length (#568)
I fixed an error in the docstring of inertial_length which still used
the definition from when it was split into ion_inertial_length and
electron_inertial_length.  I took this chance to use `particle_input`
and `Particle`, remove the default particle assumption, update
tests, and update the description of the physics.