Remove message keyword argument in pytest.raises and pytest.warns (#602)
* Remove message kwarg in pytest.raises and pytest.warns

In pytest.raises, the message argument has been used to provide a test
failure message, but it has gotten confused with the match argument
which is used to check the message from the exception to be tested so
it is being deprecated.  I changed the usage to:

    with pytest.fails(SomeException):
        should_raise_exception()
	pytest.fails("should_raise_exception() did not raise anything")

Also, we had pytest.warns(..., message="...") at one point, though it
appears pytest.warns does not do anything with the message kwarg to
begin with.  That had to be changed into a grumpier try except block.

* Improve a failure message when using pytest.warns

* Fix bug when using pytest.warns in parsing.py

* Fix bug when testing missing warnings