Remove obsolete top-level INSTALL.md file
We originally kept our installation directions in this file, but have
since moved the information to our docs directory.  We kept this file
in case it was linked to from anywhere else on the World Wide Web, but
it has been a long enough time that we can safely remove it.