Metadata updates prior to version 0.2.0 (#629)
* Remove obsolete top-level INSTALL.md file

We originally kept our installation directions in this file, but have
since moved the information to our docs directory.  We kept this file
in case it was linked to from anywhere else on the World Wide Web, but
it has been a long enough time that we can safely remove it.

* Update setup.cfg

* Update citation information

A recommended practice for scientific reproducibility is to cite the
specific version of software used for a project, and an established
practice is to cite a paper or informational resource that describes
it.  This commit updates CITATION.rst and the near-identical
docs/about/citation.rst to include both of these references.  We
previously decided to keep both of these files despite the duplication
to make it extra apparent how to cite PlasmaPy.

I changed plasmapy.__citation__ to be a list of DOI links, including
the standard reference from 2018 and version 0.2.0.  The metadata for
our releases on Zenodo could potentially change (for example, adding
the name of a contributor that wasn't known before), so a DOI is the
most robust way to go about it.

I reserved the DOI for version 0.2.0 on Zenodo, so the links to that
will not work yet.