Move pytest helpers to subdirectory of utils
Our pytest helpers functionality has been in a single file in
plasmapy.utils, which means that the functionality in there is jumbled
together with a bunch of other unrelated things.  Putting it in its
own subdirectory will help with the organization.  I also split the
existing functionality in pytest.

I updated the namespaces for run_test, etc. in a bunch of test files
throughout the package.  I also made a few non-consequential changes
in the test files that I noticed, mostly removing unused import
statements.

I am planning on doing more refactoring to the pytest helpers
functionality, which will enable me to do some more refactoring of
tests.  The API for utils.pytest_helpers should be considered unstable
at the moment.