Expand codemeta.json
I put some of the data for funding agencies in a little ahead of time,
but decided to do it now since I was able to look things up.  The NASA
HDEE and NSF CSSI proposals are currently pending.  I started working
on PlasmaPy through a DOE grant, and also did some work through an
internal Smithsonian Institution grant.