Guidelines for supporting old versions of Python, NumPy, and Astropy (#683)
* Add guidelines for supporting old versions of Python, etc.

These guidelines are adapted from NumPy Enhancement Proposal 29, which
was placed into the public domain.

* Revise discussion on supporting old Pythons & NumPys

I removed references to binary artifacts since we no longer have Cython,
and dropped the reference to testing since that issue is discussed in the
testing guide.  I clarified the discussion to allow the patch version number
of required upstream packages to be bumped during patch releases of
PlasmaPy.