Create class for expected test outcomes in pytest-helpers
The ExpectedTestOutcome class created here is an attempt to put all
responsibilities for storing and processing information about a test
outcome into one place.  This functionality has heretofore been in
our test runner function, run_test.  Because run_test contains a
whole bunch of functionality in one place, it will be much cleaner
to separate responsibilities into different classes.

I expect that this class will be used rarely outside of PlasmaPy,
so I tried to write it so that the attributes are more readable. I
don't know if this class actually should have an expected_outcome
attribute, because that is in there for the time being for human
convenience.

I will have to go back and finish up the comparison operators like
__eq__, but that will require first finishing the class that
represents that actual test outcome.  Also, we should account for
the case where the expected outcome is a unit (i.e., that the
result has units of barn-megaparsecs, for example).