Create pytest helper classes to hold test inputs
This commit creates three classes to be used to store the callable and
arguments, and then call it.  These classes include all of the
processing to handle and validate positional and keyword arguments,
and include additional checks like making sure that functions are
indeed functions.  The goal is to hide all of this low-level stuff
from the higher-level test runner functionality during the refactoring
of run_test.

The three variations of the result test pattern are: (1) calling a
function with args & kwargs, (2) calling a class instance method, and
(3) accessing a class instance attribute to make sure that the result
matches the correct value.  The classes corresponding to this include
FunctionTestInputs, ClassMethodTestInputs, and
ClassAttributeTestInputs.

The lines that say `__tracebackhide__ = True` are instructions for
pytest to consider this a test runner function and not include
traceback information when doing the error report.

An alternative to the `call` method on these classes is to use
functools.partial, which binds arguments to a callable and creates a
new function-like thing.  I think the code is more understandable to
spell out how the function, method, or attribute is being called prior
to a test.