Add first draft of a class to store outcomes of tests
I created the ActualTestOutcome class in order to call the
function/method or access a class attribute and record the outcome of
that call.  The outcome could be an exception, a warning, a returned
value, and/or both a warning and a value.  The idea is to consolidate
all of the exception handling and context environments into one place
so that it can be isolated from the high level calls.  This commit
lays out the overall structure of ActualTestOutcome, and is a good
place to look to get a big picture view before a whole bunch more
attributes and processing functionality gets added.  I wanted to make
sure to commit this now, as writing this class made my brain hurt.