Remove colorama as a dependency
We have been using colorama only in pytest_helpers to add color
formatting to test failure messages (e.g., to highlight the names of
exceptions, warnings, or the strings representing function calls).
The problem is that when the error messages are represented as plain
ASCII text, then the characters for formatting are printed out.  This
commit removes color from test failure messages and colorama as an
unnecessay dependency.