Add/update functions for call strings
I added functions that return a string that is designed to reproduce
an equivalent call.  There already was `call_string` for function
calls and class instantiation, and this commit creates two new
functions to reproduce call strings for class methods and class
attributes.  These are designed to be used in error messages.

I was wondering if these should be private functions, but they seem
potentially useful if someone else ends up creating custom error
messages.  So, I'm keeping them public for the time being.