Expand class that compares actual & expected test results
The single responsibility of the CompareActualExpected class is to
generate descriptive error messages, as necessary, for the different
reasons why tests fail.  This class accounts for cases that expect an
exception to be raised, a warning to be issued, and/or a value to be
returned.

Whereas run_test had a bunch of these things piled together in one
function, I broke up the different error message creators so that each
one has its own method in the class.  The last method in the class
(_make_errmsg_if_necessary) contains the logic of which method goes
with which conditions.  I decided against a nested structure for these
conditionals because doing otherwise started to make my brain hurt.

I still need to polish the error messages and make sure that the
correct error message matches the correct pattern.

When writing tests that fall under the result test pattern, I found
myself writing essentially the same error messages over and over for
different tests.  The goal is to isolate all of that boilerplate stuff
in this class so that we need not rewrite the same pattern over and
over.  Moreover, this allows more detailed checks to be performed.
The goal of this effort is to greatly simplify our existing test suite
as well as the process of writing new tests for situations that fall
under the result test pattern.  Once this class is completed, I should
have all of the functionality needed to rewrite run_test into
something much shorter (my goal is less than ten lines!).