Rename variables and arguments in formulary
In preparation for renaming `plasmapy.atomic` to `plasmapy.particles`,
I went through and changed variables named `particles` to `species` to
avoid the namespace conflicts that arose when I first trying the Great
Renaming.  While I was at it, I also changed `ion_particle` to `ion`
whenever it showed up.  This changes the API, in preparation for more
changes to the API.