Rename plasmapy.atomic to plasmapy.particles (#742)
* Rename variables and arguments in formulary

In preparation for renaming `plasmapy.atomic` to `plasmapy.particles`,
I went through and changed variables named `particles` to `species` to
avoid the namespace conflicts that arose when I first trying the Great
Renaming.  While I was at it, I also changed `ion_particle` to `ion`
whenever it showed up.  This changes the API, in preparation for more
changes to the API.

* Rename plasmapy.atomic to plasmapy.particles

PyCharm made this so easy!  The only thing I had to change manually
was a docstring example that imported plasmapy.atomic.  Now to check
the docstrings...  I'll be thanking Software Carpentry for teaching me
how to use grep.

* Change plasmapy.atomic to plasmapy.particles in docstrings

* Rename plasmapy.atomic to plasmapy.particles in docs

The only places where "plasmapy.atomic" should show up in the
repository are in the change log and release notes.  Those references
should stay as they are because they refer to past releases.

* Add changelog entry for plasmapy.particles renaming

* Fix typo