Update installation instructions in docs
The installation instructions before now referred to Cython and to
`pip install plasmapy[optional]`, which we no longer use.  This update
removes/updates those references and also does a little reorganization
of the docs.  I tried to simplify the installation instructions to
make it more straightforward to new users, like by dropping the SSH
cloning instructions and instead referring users to the GitHub page on
that in a note.