Set isort settings to be consistent with black (#773)
* Set isort settings to be consistent with black

The default settings for isort (which sorts imports) and black (which
strictly formats code) are not consistent with each other.  This sets
the default settings for isort to be consistent with black.  If you
run black and isort once, then switching between them afterward will
lead to no changes.

* Add changelog entry for isort defaults

* Change maximum line length to 88

This is for consistency with the default line length for the black
code formatter.