Various administrative cleanups (#790)
* Remove sunpy image-test tox-env

* Allow plot display from tests using PLASMAPY_PLOT_TESTS=1

* Add pip to environment.yml

conda nags you about not having pip as a dependency when including
pip dependencies such as numpydoc and pytest here.

* Add isort to pre-commit hooks

#773 added settings for isort, but didn't actually add isort to the suite.

* Set tox envlist to minimal set

This allows simply running `tox` to run the main package tests via py37
and the documentation, instead of taking in some cases half an hour to
run all the env combinations.

* Cleanup tox deps