Allow IonizationState to accept an ion as an argument
Up until now, the IonizationState class has required the base particle
to be either an element or isotope.  However, there may be some cases
where we want to set the ionic fraction to be 100% of a particular
ion.  This pull request allows us to do things like
IonizationState("alpha"), in which case the base particle will be He-4
and the ionic fractions will be [0.0, 0.0, 1.0] (i.e., all He-4 atoms
are doubly charged).

Credit for the idea goes to:
Co-authored-by: Maurice Wilson <maurice.wilson@cfa.harvard.edu>