Create first attempt at abstract base class for normalizations
I expect that an abstract base class will be helpful here because will be
different normalizations for different ways of representing plasma (e.g,
ideal, resistive MHD, or extended MHD normalizations are different than
normalizations used for PIC simulations.  I'll tentatively keep this in,
but we can re-evaluate how useful it will be later on.