Adjust RST heading lengths
This kind of thing is why Markdown is the superior markup language.

Did you know Julia has Markdown as a default language for docs?

It's pretty cool, it's as if they looked at everything that works in
Python and decided to make it better.

Fanboying so hard right now, wow.