Replace instances of Riot with Matrix/Element (#891)
* Replace instances of Riot with Matrix/Element

To quote from the relevant issue:

The interesting thing here:

* Matrix is the chat platform we were and still are using;
* Riot is the old name of the particular chat client we're using;
* Element is the new name for that chat client.

> nowadays there are lots of other Matrix clients, vendors and hosting providers that demonstrate perfectly the diversity of the Matrix ecosystem.

(from the linked page)

So I'm going to standardize this to Matrix in most cases, but link to Element. I guess we could link to https://matrix.to/#/#plasmapy:openastronomy.org (I found the site by looking at what provides the ![Matrix](https://matrix.to/img/matrix-badge.svg) badge), but since Matrix is a bit niche as is, having the chat another step away and making people decide which chat client to use seems counterproductive.

* Changelog entry

* Adjust RST heading lengths

This kind of thing is why Markdown is the superior markup language.

Did you know Julia has Markdown as a default language for docs?

It's pretty cool, it's as if they looked at everything that works in
Python and decided to make it better.

Fanboying so hard right now, wow.