Update citation recommendations and release guide (#900)
* Adopt docs/citation.rst as the main place for citation info

There are currently several places where we have information on how to
cite PlasmaPy, and some of them are out of date.  This is to use the
version in the documentation as the main version.

* Minor updates to release guide

* Begin updates for the release guide regarding citations

* Add changelog entry for #900

* Update information on how to cite PlasmaPy

* Link to citation page in stable version of docs

The version that someone is most likely to cite is the stable version
that has already had a release, rather than the latest version.  This
does presume that we update docs/about/citation.rst, but this task is
being added to the release guide.

* Fix link to citation file in README file