Refactor _code_repr_of_arg to remove extended conditional block
This creates a dict with types as keys, and code representation functions as
values.  The function goes through the types, and checks if the argument is an
instance of that type.  If so, it selects the appropriate code representation
function.  If all else fails, then it uses repr, which does exactly what is
needed for str, int, etc. objects.