Update function to get indefinite article before exception name
The function to get the indefinite article is fine-tuned for exceptions
and will not handle unusual cases beyond that.  We could potentially use
a library like `inflect` on PyPI since it seems to handle cases where a
word starts with a vowel but uses "a", but it's not worth adding an
extra dependency yet.