Update pickled particles tests
The open statement needed to have a file access mode of "rb" rather than
"r" in order to be unpickled.  This also compares string representations
of the original and loaded objects rather than the objects themselves
because __eq__ might not be implemented.  The comparison mechanism could
be expanded later on.