Add mostly failing tests for pickling various objects in plasmapy.particles (#1014)
* Add tests for pickled particles

* Add failing tests for pickling ionization state classes

* Consolidate tests of particle paraphernalia pickling

* Replace teardown_method with tmp_path

see https://docs.pytest.org/en/stable/tmpdir.html

* Add logical next test to check that particle can actually be loaded back

* Update pickled particles tests

The open statement needed to have a file access mode of "rb" rather than
"r" in order to be unpickled.  This also compares string representations
of the original and loaded objects rather than the objects themselves
because __eq__ might not be implemented.  The comparison mechanism could
be expanded later on.

Co-authored-by: Dominik Stańczak <stanczakdominik@gmail.com>