Remove caching in ParticleList (+ other updates)
Suppose one element of ParticleList is a CustomParticle. Then suppose
someone modifies the mass of that CustomParticle directly, without
modifying the identities of the particles within the list. Then the
cached values of mass would no longer be valid. This will probably not
be a common issue, but it could cause unexpected and difficult to track
down problems. It'll be safer and cleaner to avoid caching. We can
revisit caching if/when there's a situation where it ends up being needed.