Create custom list for particles (#1017)
* Create file for ParticleList

* Begin ParticleList and tests

* Expand ParticleList

* Add antiparticle attributes to CustomParticle & DimensionlessParticle

* Create first round of tests for ParticleList

* Add incomplete draft of ParticleList

* Update ParticleList and try out a class specifically for ions.

* Update tests of ParticleList

They're not passing right now, but that seems to be a problem with like 61 of the
test file.

* Add a particle attribute for CustomParticle

* Update ParticleList

* Alphabetize attributes

* Remove caching in ParticleList (+ other updates)

Suppose one element of ParticleList is a CustomParticle. Then suppose
someone modifies the mass of that CustomParticle directly, without
modifying the identities of the particles within the list. Then the
cached values of mass would no longer be valid. This will probably not
be a common issue, but it could cause unexpected and difficult to track
down problems. It'll be safer and cleaner to avoid caching. We can
revisit caching if/when there's a situation where it ends up being needed.

* Update ParticleList tests

The tests still need to be cleaned up and expanded more.

* Revise ParticleLike

* Modify tests for ParticleLike

* Finish changing particle_like to ParticleLike

* Add changelog entry for ParticleLike

* Test list-like behavior for ParticleList

* Refine tests & docstrings for ParticleList

* Use + and > with Particles & ParticleLists to get nuclear reaction energy

I believe this idea came from Stuart Mumford several years ago.

* Follow-up on previous commit

* Expand and clean up ParticleList tests

* Minor clean up of ParticleList

* Add TODO note about issue #687

Co-authored-by: Dominik Stańczak <stanczakdominik@gmail.com>

* Use generator rather than iterator

Co-authored-by: Dominik Stańczak <stanczakdominik@gmail.com>

* Rename nuclear test and add regression test value

* Expand docstring for ParticleList

* Create AbstractPhysicalParticle class and update docstring for Particle

* Add docstring

* Expand tests for ParticleList

* Expand changelog for ParticleList

* Add docstrings for methods in `ParticleList`

These were not showing up correctly in the rendered docs.

* Refine docstrings for ParticleList methods

* Remove extra lines from nuclear_reaction_energy docstring

* Minor revisions to docstrings

* Attempt to get ParticleLike to render better in documentation

Right now it shows up as typing.Union[blah].

* Fix the earlier fixes to Particle docstring

* Include imports of abstract base classes

* Wording fix

Co-authored-by: Dominik Stańczak <stanczakdominik@gmail.com>

* Update docstrings and comments

* Tentatively remove ParticleLike.__name__

I'm wondering if this might be causing doc build errors related to finding
`ParticleLike`.

Co-authored-by: Dominik Stańczak <stanczakdominik@gmail.com>