Add looponfailroots=plasmapy to setup.cfg
This allows running tests with pytest-xdist's `pytest --looponfail`
setting. What that does is run tests in a loop whenever a file is
modified,, but if some fail, it limits the next loop to those particular
tests. It's awesome.  Anyway, looponfailroots limits the file modified
detection to the plasmapy dir, which is reasonable here. I had it
trigger on vim creating those tags files I mentioned earlier.