Update paschen.py
I tried my best to correct the paschen.py. I was reading many tutorials and examples but at the end it is difficult to remember all the different norms (ufff). Sorry there might be still many errors, sorry to give you soo much work, but this part is for a beginner much more difficult to accomplish than writing the program itself. However I understand that it must be like this to have correct contributions.
I let some print for testing
I am trying to add the units, still fight on this.
I send the present version to advance a little bit (sorry I was busy with different important and less important things...)