Add an interface function that creates physical particle objects (#1365)
* Continue working on particle factory & tests

* Move particle factory function to a new file

This is to avoid having to avoid circularity. particle_class.py
defines Particle, and particle_collections.py imports Particle to
define ParticleList. Prior to this commit, we had the factory function
in particle_class.py, where it imported ParticleList.

* Allow first two arguments of CustomParticle to be switched

* Finish CustomParticle having reversed argument order

* Update particle factory

* Remove edits to CustomParticle

These edits were moved to a different pull request.

* Add changelog entry

* Simplify discussion of parameters

* Mark a test as xfail

This will be ex-xfailed after #1364 is merged.

* Temporarily add particle factory stuff to __all__

* Make particle factory function private

* Refine docs & tests of particle factory