Change release to 0.8.1
I previously created the v0.8.0 tag, but the release step errored out
because there was a problem with PyPI authentication. Unfortunately,
once a tag is created on GitHub, it can't be deleted (‽‽‽). Because
releases are based on tags, I don't appear to be able to create a
v0.8.0 release.  The least headacheful workaround is to skip v0.8.0
and release v0.8.1 instead.