Merge `v0.8.x` branch back into `main` (#1616)
* Run towncrier

* Add docs/whatsnew/0.8.0.rst

* Change capitalization

* Update release notes for 0.8.0

* Fix capitalization in PlasmaPy

* Fix spelling in name

* Fix spelling in name

* Update GitHub Action for publishing to PyPI

* Remove blank line

* Make build-n-publish depend on all other ci checks

* Update dates in whatsnew entry

* Add python-publish.yml workflow

* Temporarily remove action to publish upon pushing a tagged commit

* Update date of release

* Be more specific about GA version

* Slight edits to whatsnew page

* Change release to 0.8.1

I previously created the v0.8.0 tag, but the release step errored out
because there was a problem with PyPI authentication. Unfortunately,
once a tag is created on GitHub, it can't be deleted (‽‽‽). Because
releases are based on tags, I don't appear to be able to create a
v0.8.0 release.  The least headacheful workaround is to skip v0.8.0
and release v0.8.1 instead.

* Run pre-commit --all