[TEMPORARY] do [docs] need [tests]?
Based on https://github.com/PlasmaPy/PlasmaPy/actions/runs/3039593035/jobs/4894695172#step:6:22, it looks like the docs environment is failing because it can't import `plasmapy.utils.pytest_helpers`. If so, I guess we should include `pytest` as a dependency of the [docs] extra. Hell, we could even do, in `requirements/docs.txt`,

```
sphinx
pyramid
hieroglyph
...
-r tests.txt
```

or some such, so that `tests` is strictly a subset of `docs`. But getting just pytest in is probably faster on the CI executor.