Change up GitHub contributing file
Previously, we had the corresponding file in `docs/CONTRIBUTING.rst`.
Unfortunately, that did not get linked to correctly from GitHub since
it had to be a Markdown file, so I changed it to a Markdown file. I
also moved it out of `docs/` since it is a Markdown file, and into
`.github` since it is intended for GitHub usage and so that it does not
add extra clutter to our top-level directory. I've also simplified it
to avoid duplication of content, and instead have it link to our
contributor guide and code of conduct.