Add mastodon badge to readme (#1895)
This adds our mastodon account to our README. The `<a rel="me">` part serves as a verification mechanism - if you add a link like that to a $webpage, and link from the account to $webpage, it checks whether $webpage contains that link and displays a checkmark if it does. It's neat.