Remove nbqa-ruff config settings from pyproject.toml
I had trouble getting nbqa-ruff to use the settings that were in pyproject.toml, so
I opted to move the settings to .pre-commit-config.yaml instead. The downside
is that there are two different rule sets for ruff and nbqa-ruff...but this should be
okay since we want to apply different rules (i.e., we want print statements in
notebooks but usually not in production code).