Update sphinx-build flags for on/off RTD
This change makes it so that different flags get used for sphinx-build
depending on whether the Nox session is run on RTD or off of it.

Usually it is easier to parse the standard output when the `--quiet`
flag is used, so it is the default. Since it can be helpful to see the
full output, the RTD build does not use the `--quiet` flag.

The `--fail-on-warning` flag is not used on RTD because it makes it
harder to see the documentation preview, which can be really helpful.