Remove fetch-depth for actions/checkout
When checking out a GitHub repository in a workflow, we had been setting
fetch-depth=0 because we intended to only include the most recent commit.
However, fetch-depth=0 causes the entire history to be downloaded. Oops.
We should have had fetch-depth=1 since that downloads only the most
recent commit, but since that's the default, we do not need to set
fetch-depth at all.