Ignore some type checking errors related to validate_quantities
Currently, the @validate_quantities decorator remains to be fully typed, so
there are a few errors that arise when it is used in different modules. This
commit updates the mypy configuration to ignore a few static type checking
errors in the new module and test file created in this pull request.