Reduce warnings in tests (#3074)
* Planned removal of Particle.binding_energy

* Remove Particle.binding_energy test

* Update pytest durations setting

* Change kinetic_alfven example to not emit warning

* Adjust kinetic_alfven example to avoid warnings

* Hide ion_sound_speed warning from within kinetic_alfven()

My reasoning for ignoring this warning is that the warnings
later on check for the validity of this dispersion relation.

* Provide numbers in warning messages

* Update warning messages for hollweg()

* spacing

* Fix import statement

* Update Hollweg example to not emit warning

* binding_energy -> nuclear_binding_energy

* Edit changelog: 3074.removal.rst

* Edit changelog: 3074.removal.rst

* Update warnings filters in formulary

* Ignore warnings in test_pytest_helpers

* Update pytest durations options

* Update pytest filterwarnings in pyproject.toml

* Apply @filterwarnings in test_coulomb.py

* Mark downloader tests as slow

While not strictly slow, the downloader tests can complicate things
if an internet connection is unavailable or multiple API calls need
to be made.

* Update warnings handling in particle tracker tests

* Fix rst link

* Edit changelog: 3074.internal.rst

* Remove duplicate decorators after merge

* Change --durations-min

* Update comments about filterwarnings

* Decorate with @pytest.mark.filterwarnings

* Ignore some weird GitHub warnings

* Update changelog entries

* Allow warnings when testing against lowest-direct dependencies

This is one case where I measured the variable name length not
by the number of characters, but by the time needed to understand
its meaning and intention.

* codespell

* Remove CI doc build for Python 3.14

This was temporary anyway, and it wasn't working because the session in noxfile.py assumed that it was Python 3.13 anyway so it wasn't working.