Remove monkeytype session for Nox
This session did not work for me just now when I tried it. Because it's
a complicated session that will be rare for us to use, I'm going to
remove it. Also, pytest-monkeytype seems to not have been updated in
the last several years, so I'm not sure if it'll remain interoperable
with pytest.