Use @nox_uv.session in docs_upstream
This session previously took the names optional dependencies from
pyproject.toml without the version specifiers and passed them into
session.install(). The motivation was to avoid resolution conflicts
if an upper dependency limit had been put on the target package.
I believe that this is no longer necessary when using @nox_uv.session.
At the same time, it is helpful information to know that there is a
resolution conflict that will need to be fixed, and failing when that
happens will inform us about that problem.